Move related frame instantiation and drawing methods from Play into MyFrame
SnakeObjects such as Food and MySnake instantiations, background image fetches, drawScore and draw are moved into MyFrame so that Play class has only one responsibility which is to launch the game.

MyFrame class would be responsible to draw or set up the frame and scene.

This refactoring obeys Single Responsibility Principle and loosens the coupling between Play class and MyFrame class.